impl store seating functionality

+++ internal/account/contract.go
@@ -19,7 +19,7 @@ type IAccountRepository interface {
 type IAccountService interface {
 	Roles(ctx context.Context) ([]*model.Role, *utils.ServiceError)
 	Users(ctx context.Context) ([]*model.User, *utils.ServiceError)
-	UserByID(ctx context.Context, id int) (*model.User, *utils.ServiceError)
+	UserByID(ctx context.Context, id int64) (*model.User, *utils.ServiceError)
 	CreateUser(ctx context.Context, data *NewUserForm) (*model.User, *utils.ServiceError)
 	UpdateUser(ctx context.Context, data *UpdateUserForm) (*model.User, *utils.ServiceError)
 	UpdateUserPassword(ctx context.Context, data *UpdatePasswordForm) *utils.ServiceError
